git-svn-id: http://springhead.info/spr2/Springhead2/trunk@1709 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHSolid.cpp
@@ -236,13 +236,13 @@ void PHSolid::AddShape(CDShapeIf* shape){
 }
 
 Posed	PHSolid::GetShapePose(int i){
-	if(0 <= i && i < shapes.size())
+	if(0 <= i && i < (int)shapes.size())
 		return shapes[i].pose;
 	return Posed();
 }
 
 void	PHSolid::SetShapePose(int i, const Posed& pose){
-	if(0 <= i && i < shapes.size())
+	if(0 <= i && i < (int)shapes.size())
 		shapes[i].pose = pose;
 }
